👌 IMPROVE: Improve course buttons
