+++ src/components/CourseFinder/__mock__/CoursesCodeTitleCredits.json
