+++ src/components/CourseFinder/__mock__/CoursesNoDescription.json
